set relation for track to audiofile
