+++ AudioCuesheetEditor/Model/AudioCuesheet/Cuesheet.cs
@@ -21,10 +21,6 @@ namespace AudioCuesheetEditor.Model.AudioCuesheet
 {
     public class Cuesheet() : Validateable, ICuesheet
     {
-        //TODO: Move tracks below audiofiles
-        [JsonInclude]
-        public IEnumerable<Track> Tracks { get; set; } = [];
-
         public String? Artist { get; set; }
         
         public String? Title { get; set; }
@@ -47,46 +43,6 @@ namespace AudioCuesheetEditor.Model.AudioCuesheet
             List<ValidationMessage>? validationMessages = null;
             switch (property)
             {
-                case nameof(Tracks):
-                    validationStatus = ValidationStatus.Success;
-                    if (!Tracks.Any())
-                    {
-                        validationMessages ??= [];
-                        validationMessages.Add(new ValidationMessage("{0} has invalid count ({1})!", nameof(Tracks), 0));
-                    }
-                    else
-                    {
-                        //Check track overlapping
-                        var tracksWithSamePosition = Tracks
-                            .GroupBy(x => x.Position)
-                            .Where(grp => grp.Count() > 1);
-                        if (tracksWithSamePosition.Any())
-                        {
-                            validationMessages ??= [];
-                            foreach (var track in tracksWithSamePosition)
-                            {
-                                foreach (var trackWithSamePosition in track)
-                                {
-                                    validationMessages.Add(new ValidationMessage("{0} {1} '{2}' is used also by {3}({4},{5},{6},{7},{8}). Positions must be unique!", nameof(Track), nameof(Track.Position), track.Key != null ? track.Key : String.Empty, nameof(Track), trackWithSamePosition.Position != null ? trackWithSamePosition.Position : String.Empty, trackWithSamePosition.Artist ?? String.Empty, trackWithSamePosition.Title ?? String.Empty, trackWithSamePosition.Begin != null ? trackWithSamePosition.Begin : String.Empty, trackWithSamePosition.End != null ? trackWithSamePosition.End : String.Empty));
-                                }
-                            }
-                        }
-                        foreach (var track in Tracks.OrderBy(x => x.Position))
-                        {
-                            var tracksBetween = Tracks.Where(x => ((track.Begin >= x.Begin && track.Begin < x.End)
-                                                        || (x.Begin < track.End && track.End <= x.End))
-                                                        && (x.Equals(track) == false));
-                            if (tracksBetween.Any())
-                            {
-                                validationMessages ??= [];
-                                foreach (var trackBetween in tracksBetween)
-                                {
-                                    validationMessages.Add(new ValidationMessage("{0}({1},{2},{3},{4},{5}) is overlapping with {0}({6},{7},{8},{9},{10}). Please make shure the timeinterval is only used once!", nameof(Track), track.Position != null ? track.Position : String.Empty, track.Artist ?? String.Empty, track.Title ?? String.Empty, track.Begin != null ? track.Begin : String.Empty, track.End != null ? track.End : String.Empty, trackBetween.Position != null ? trackBetween.Position : String.Empty, trackBetween.Artist ?? String.Empty, trackBetween.Title ?? String.Empty, trackBetween.Begin != null ? trackBetween.Begin : String.Empty, trackBetween.End != null ? trackBetween.End : String.Empty));
-                                }
-                            }
-                        }
-                    }
-                    break;
                 case nameof(Audiofiles):
                     validationStatus = ValidationStatus.Success;
                     if (Audiofiles.Count == 0)
