+++ AudioCuesheetEditor/Model/AudioCuesheet/Track.cs
@@ -67,8 +67,6 @@ namespace AudioCuesheetEditor.Model.AudioCuesheet
         }
         [JsonInclude]
         public IEnumerable<Flag> Flags { get; set; } = [];
-        [JsonIgnore]
-        public Cuesheet? Cuesheet { get; set; }
         /// <inheritdoc/>
         public TimeSpan? PreGap { get; set;  }
         /// <inheritdoc/>
@@ -101,15 +99,16 @@ namespace AudioCuesheetEditor.Model.AudioCuesheet
                         else
                         {
                             // Check correct track position
-                            if (Cuesheet != null)
-                            {
-                                var positionTrackShouldHave = Cuesheet.Tracks.OrderBy(x => x.Begin ?? TimeSpan.MaxValue).ThenBy(x => x.Position).ToList().IndexOf(this) + 1;
-                                if (positionTrackShouldHave != Position)
-                                {
-                                    validationMessages ??= [];
-                                    validationMessages.Add(new ValidationMessage("Track({0},{1},{2},{3},{4}) does not have the correct position '{5}'!", Position, Artist ?? String.Empty, Title ?? String.Empty, Begin != null ? Begin : String.Empty, End != null ? End : String.Empty, positionTrackShouldHave));
-                                }
-                            }
+                            //TODO
+                            //if (Cuesheet != null)
+                            //{
+                            //    var positionTrackShouldHave = Cuesheet.Tracks.OrderBy(x => x.Begin ?? TimeSpan.MaxValue).ThenBy(x => x.Position).ToList().IndexOf(this) + 1;
+                            //    if (positionTrackShouldHave != Position)
+                            //    {
+                            //        validationMessages ??= [];
+                            //        validationMessages.Add(new ValidationMessage("Track({0},{1},{2},{3},{4}) does not have the correct position '{5}'!", Position, Artist ?? String.Empty, Title ?? String.Empty, Begin != null ? Begin : String.Empty, End != null ? End : String.Empty, positionTrackShouldHave));
+                            //    }
+                            //}
                         }
                     }
                     break;
