+++ AudioCuesheetEditor/Services/AudioCuesheet/AudiofileManager.cs
@@ -13,6 +13,7 @@
 //You should have received a copy of the GNU General Public License
 //along with Foobar.  If not, see
 //<http: //www.gnu.org/licenses />.
+using AudioCuesheetEditor.Model.AudioCuesheet;
 using AudioCuesheetEditor.Model.IO.Audio;
 using AudioCuesheetEditor.Services.IO;
 using AudioCuesheetEditor.Services.UI;
@@ -23,11 +24,13 @@ using System.Reflection;
 
 namespace AudioCuesheetEditor.Services.AudioCuesheet
 {
-    public class AudiofileManager(IFileInputManager fileInputManager, ITraceChangeManager traceChangeManager, IJSRuntime jsRuntime) : IAudiofileManager
+    public class AudiofileManager(IFileInputManager fileInputManager, ITraceChangeManager traceChangeManager, IJSRuntime jsRuntime, ITrackManager trackManager, ISessionStateContainer sessionStateContainer) : IAudiofileManager
     {
         private readonly IFileInputManager _fileInputManager = fileInputManager;
         private readonly ITraceChangeManager _traceChangeManager = traceChangeManager;
         private readonly IJSRuntime _jsRuntime = jsRuntime;
+        private readonly ITrackManager _trackManager = trackManager;
+        private readonly ISessionStateContainer _sessionStateContainer = sessionStateContainer;
 
         /// <inheritdoc/>
         public async Task SetPropertiesAsync(Audiofile audiofile, IBrowserFile? browserFile, string fileInputId)
@@ -59,10 +62,67 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
             _traceChangeManager.BulkEdit = false;
         }
 
+        /// <inheritdoc/>
         public void SetProperty<TProperty>(Audiofile audiofile, Expression<Func<Audiofile, TProperty>> propertyExpression, TProperty value)
         {
             //TODO: Tests
             SetValue(audiofile, propertyExpression, value);
+        }
+
+        /// <inheritdoc/>
+        public void AddTrack(Audiofile audiofile, Track track)
+        {
+            //TODO: Tests
+            //Calculate track properties
+            _traceChangeManager.BulkEdit = true;
+            var cuesheet = _sessionStateContainer.GetActiveCuesheet();
+            if (cuesheet?.IsRecording == true)
+            {
+                _trackManager.SetProperty(track, x => x.Begin, DateTime.UtcNow - cuesheet.RecordingStart);
+            }
+            if (cuesheet?.Audiofiles.SelectMany(x => x.Tracks).Any() == false)
+            {
+                _trackManager.SetProperty(track, x => x.Position, (ushort)(1));
+                if ((track.Begin.HasValue == false) || cuesheet?.IsRecording == true)
+                {
+                    _trackManager.SetProperty(track, x => x.Begin, TimeSpan.Zero);
+                }
+            }
+            else
+            {
+                //TODO: adapt to multiple audio files
+                var lastTrack = GetLastTrack(cuesheet!);
+                //if ((cuesheet?.Audiofile?.Duration.HasValue == true) && (lastTrack?.End.HasValue == true) && (lastTrack.End == cuesheet.Audiofile.Duration))
+                //{
+                //    _trackManager.SetProperty(lastTrack, x => x.End, null);
+                //}
+                if (track.Position.HasValue == false)
+                {
+                    _trackManager.SetProperty(track, x => x.Position, (ushort?)(lastTrack?.Position + 1));
+                }
+                if (track.Begin.HasValue == false)
+                {
+                    _trackManager.SetProperty(track, x => x.Begin, lastTrack?.End);
+                }
+                else
+                {
+                    if (lastTrack?.End.HasValue == false)
+                    {
+                        _trackManager.SetProperty(lastTrack, x => x.End, track.Begin);
+                    }
+                }
+                if (cuesheet?.IsRecording == true && lastTrack != null)
+                {
+                    _trackManager.SetProperty(lastTrack, x => x.End, track.Begin);
+                }
+            }
+            var newValue = new List<Track>(audiofile.Tracks)
+            {
+                track
+            };
+            SetValue(audiofile, x => x.Tracks, newValue);
+            SetLastTrackEnd(cuesheet!);
+            _traceChangeManager.BulkEdit = false;
         }
 
         void SetValue<TProperty>(Audiofile audiofile, Expression<Func<Audiofile, TProperty>> propertyExpression, TProperty value)
@@ -85,6 +145,25 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
 
             propertyInfo.SetValue(audiofile, value);
             _traceChangeManager.AddChange(new(audiofile, new(previousValue, propertyInfo.Name)));
+        }
+
+        void SetLastTrackEnd(Cuesheet cuesheet)
+        {
+            var lastTrack = GetLastTrack(cuesheet);
+            //TODO
+            //if ((lastTrack?.End.HasValue == false) && (cuesheet.Audiofile?.Duration.HasValue == true))
+            //{
+            //    _trackManager.SetProperty(lastTrack, x => x.End, cuesheet.Audiofile.Duration);
+            //}
+        }
+
+        static Track? GetLastTrack(Cuesheet cuesheet)
+        {
+            return cuesheet.Audiofiles.SelectMany(x => x.Tracks)
+                .OrderByDescending(x => x.Position.HasValue).ThenBy(x => x.Position)
+                .ThenByDescending(x => x.Begin.HasValue).ThenBy(x => x.Begin)
+                .ThenByDescending(x => x.End.HasValue).ThenBy(x => x.End)
+                .LastOrDefault();
         }
     }
 }
