+++ AudioCuesheetEditor/Services/AudioCuesheet/CuesheetManager.cs
@@ -50,10 +50,11 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
         {
             get
             {
-                if (_sessionStateContainer.Cuesheet.Tracks.Any())
-                {
-                    return Result.Failure(new Error(ErrorType.NotPossible, "Cuesheet already contains tracks!"));
-                }
+                //TODO
+                //if (_sessionStateContainer.Cuesheet.Tracks.Any())
+                //{
+                //    return Result.Failure(new Error(ErrorType.NotPossible, "Cuesheet already contains tracks!"));
+                //}
                 return Result.Success();
             }
         }
@@ -82,101 +83,47 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
             var cuesheet = _sessionStateContainer.Cuesheet;
             if (cuesheet.IsRecording == true)
             {
-                var lastTrack = cuesheet.Tracks.LastOrDefault();
-                if ((lastTrack != null) && cuesheet.RecordingStart.HasValue)
-                {
-                    lastTrack.End = DateTime.UtcNow - cuesheet.RecordingStart.Value;
-                }
+                //TODO
+                //var lastTrack = cuesheet.Tracks.LastOrDefault();
+                //if ((lastTrack != null) && cuesheet.RecordingStart.HasValue)
+                //{
+                //    lastTrack.End = DateTime.UtcNow - cuesheet.RecordingStart.Value;
+                //}
                 cuesheet.RecordingStart = null;
                 IsRecordingChanged?.Invoke(this, EventArgs.Empty);
             }
         }
 
         /// <inheritdoc/>
-        public void AddTrack(Track track)
-        {
-            var cuesheet = _sessionStateContainer.GetActiveCuesheet();
-            track.Cuesheet = cuesheet;
-            // Calculate track properties
-            _traceChangeManager.BulkEdit = true;
-            if (cuesheet?.IsRecording == true)
-            {
-                _trackManager.SetProperty(track, x => x.Begin, DateTime.UtcNow - cuesheet.RecordingStart);
-            }
-            if (cuesheet?.Tracks.Any() == false)
-            {
-                _trackManager.SetProperty(track, x => x.Position, (ushort)(1));
-                if ((track.Begin.HasValue == false) || cuesheet.IsRecording)
-                {
-                    _trackManager.SetProperty(track, x => x.Begin, TimeSpan.Zero);
-                }
-            }
-            else
-            {
-                var lastTrack = GetLastTrack(cuesheet!);
-                //TODO
-                //if ((cuesheet?.Audiofile?.Duration.HasValue == true) && (lastTrack?.End.HasValue == true) && (lastTrack.End == cuesheet.Audiofile.Duration))
-                //{
-                //    _trackManager.SetProperty(lastTrack, x => x.End, null);
-                //}
-                if (track.Position.HasValue == false)
-                {
-                    _trackManager.SetProperty(track, x => x.Position, (ushort?)(lastTrack?.Position + 1));
-                }
-                if (track.Begin.HasValue == false)
-                {
-                    _trackManager.SetProperty(track, x => x.Begin, lastTrack?.End);
-                }
-                else
-                {
-                    if (lastTrack?.End.HasValue == false)
-                    {
-                        _trackManager.SetProperty(lastTrack, x => x.End, track.Begin);
-                    }
-                }
-                if (cuesheet?.IsRecording == true && lastTrack != null)
-                {
-                    _trackManager.SetProperty(lastTrack, x => x.End, track.Begin);
-                }
-            }
-            var newValue = new List<Track>(cuesheet!.Tracks)
-            {
-                track
-            };
-            SetValue(cuesheet, x => x.Tracks, newValue);
-            SetLastTrackEnd(cuesheet);
-            _traceChangeManager.BulkEdit = false;
-        }
-
-        /// <inheritdoc/>
         public void RemoveTracks(IEnumerable<Track> tracksToRemove)
         {
             var cuesheet = _sessionStateContainer.GetActiveCuesheet();
-            var intersection = cuesheet!.Tracks.Intersect(tracksToRemove);
-            ICollection<Track> newValue = [.. cuesheet.Tracks.Except(intersection)];
-            //Calculate position and begin of new tracks
-            ushort position = 1;
-            foreach (var track in newValue.OrderBy(x => x.Position))
-            {
-                track.Position = position;
-                position++;
-                var previousTrack = _trackManager.GetPreviousLinkedTrack(track);
-                if (previousTrack?.End.HasValue == true)
-                {
-                    track.Begin = previousTrack.End;
-                }
-            }
-            _traceChangeManager.BulkEdit = true;
-            SetValue(cuesheet, x => x.Tracks, newValue);
-            SetLastTrackEnd(cuesheet);
-            _traceChangeManager.BulkEdit = false;
+            //TODO
+            //var intersection = cuesheet!.Tracks.Intersect(tracksToRemove);
+            //ICollection<Track> newValue = [.. cuesheet.Tracks.Except(intersection)];
+            ////Calculate position and begin of new tracks
+            //ushort position = 1;
+            //foreach (var track in newValue.OrderBy(x => x.Position))
+            //{
+            //    track.Position = position;
+            //    position++;
+            //    var previousTrack = _trackManager.GetPreviousLinkedTrack(track);
+            //    if (previousTrack?.End.HasValue == true)
+            //    {
+            //        track.Begin = previousTrack.End;
+            //    }
+            //}
+            //_traceChangeManager.BulkEdit = true;
+            //SetValue(cuesheet, x => x.Tracks, newValue);
+            //SetLastTrackEnd(cuesheet);
+            //_traceChangeManager.BulkEdit = false;
         }
 
         /// <inheritdoc/>
         public bool IsMoveTracksUpPossible(HashSet<Track> selectedTracks) => selectedTracks.Count > 0 && selectedTracks.Min(x => x.Position) >= 2;
 
         /// <inheritdoc/>
-        public bool IsMoveTracksDownPossible(HashSet<Track> selectedTracks) => selectedTracks.Count > 0 && selectedTracks.Max(x => x.Position) < _sessionStateContainer.GetActiveCuesheet()?.Tracks.Max(x => x.Position);
+        public bool IsMoveTracksDownPossible(HashSet<Track> selectedTracks) => selectedTracks.Count > 0; //TODO &&  selectedTracks.Max(x => x.Position) < _sessionStateContainer.GetActiveCuesheet()?.Tracks.Max(x => x.Position);
 
         /// <inheritdoc/>
         public Result MoveTracksUp(HashSet<Track> selectedTracks)
@@ -187,22 +134,23 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
             }
             _traceChangeManager.BulkEdit = true;
             var cuesheet = _sessionStateContainer.GetActiveCuesheet();
-            foreach (var selectedTrack in selectedTracks.OrderBy(x => x.Position))
-            {
-                var previousTrack = cuesheet?.Tracks.FirstOrDefault(x => x.Position == selectedTrack.Position - 1);
-                var newBegin = previousTrack?.Begin;
-                var newEnd = previousTrack?.End;
-                if (previousTrack != null)
-                {
-                    _trackManager.SetProperty(previousTrack, x => x.Position, selectedTrack.Position);
-                    _trackManager.SetProperty(previousTrack, x => x.Begin, selectedTrack.Begin);
-                    _trackManager.SetProperty(previousTrack, x => x.End, selectedTrack.End);
-                }
-                _trackManager.SetProperty(selectedTrack, x => x.Position, (ushort?)(selectedTrack.Position - 1));
-                _trackManager.SetProperty(selectedTrack, x => x.Begin, newBegin);
-                _trackManager.SetProperty(selectedTrack, x => x.End, newEnd);
-            }
-            SetValue(cuesheet!, x => x.Tracks, cuesheet?.Tracks.OrderBy(x => x.Position));
+            //TODO
+            //foreach (var selectedTrack in selectedTracks.OrderBy(x => x.Position))
+            //{
+            //    var previousTrack = cuesheet?.Tracks.FirstOrDefault(x => x.Position == selectedTrack.Position - 1);
+            //    var newBegin = previousTrack?.Begin;
+            //    var newEnd = previousTrack?.End;
+            //    if (previousTrack != null)
+            //    {
+            //        _trackManager.SetProperty(previousTrack, x => x.Position, selectedTrack.Position);
+            //        _trackManager.SetProperty(previousTrack, x => x.Begin, selectedTrack.Begin);
+            //        _trackManager.SetProperty(previousTrack, x => x.End, selectedTrack.End);
+            //    }
+            //    _trackManager.SetProperty(selectedTrack, x => x.Position, (ushort?)(selectedTrack.Position - 1));
+            //    _trackManager.SetProperty(selectedTrack, x => x.Begin, newBegin);
+            //    _trackManager.SetProperty(selectedTrack, x => x.End, newEnd);
+            //}
+            //SetValue(cuesheet!, x => x.Tracks, cuesheet?.Tracks.OrderBy(x => x.Position));
             _traceChangeManager.BulkEdit = false;
             return Result.Success();
         }
@@ -215,25 +163,26 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
             {
                 return Result.Failure(new Error(ErrorType.NotPossible, "Moving tracks down is not possible!"));
             }
-            _traceChangeManager.BulkEdit = true;
-            foreach (var selectedTrack in selectedTracks.OrderByDescending(x => x.Position))
-            {
-                var nextTrack = cuesheet?.Tracks.FirstOrDefault(x => x.Position == selectedTrack.Position + 1);
-                var newBegin = nextTrack?.Begin;
-                var newEnd = nextTrack?.End;
-                if (nextTrack != null)
-                {
-                    _trackManager.SetProperty(nextTrack, x => x.Position, selectedTrack.Position);
-                    _trackManager.SetProperty(nextTrack, x => x.Begin, selectedTrack.Begin);
-                    _trackManager.SetProperty(nextTrack, x => x.End, selectedTrack.End);
-                }
-                var newPosition = (ushort?)(selectedTrack.Position + 1);
-                _trackManager.SetProperty(selectedTrack, x => x.Position, newPosition);
-                _trackManager.SetProperty(selectedTrack, x => x.Begin, newBegin);
-                _trackManager.SetProperty(selectedTrack, x => x.End, newEnd);
-            }
-            SetValue(cuesheet!, x => x.Tracks, cuesheet?.Tracks.OrderBy(x => x.Position));
-            _traceChangeManager.BulkEdit = false;
+            //TODO
+            //_traceChangeManager.BulkEdit = true;
+            //foreach (var selectedTrack in selectedTracks.OrderByDescending(x => x.Position))
+            //{
+            //    var nextTrack = cuesheet?.Tracks.FirstOrDefault(x => x.Position == selectedTrack.Position + 1);
+            //    var newBegin = nextTrack?.Begin;
+            //    var newEnd = nextTrack?.End;
+            //    if (nextTrack != null)
+            //    {
+            //        _trackManager.SetProperty(nextTrack, x => x.Position, selectedTrack.Position);
+            //        _trackManager.SetProperty(nextTrack, x => x.Begin, selectedTrack.Begin);
+            //        _trackManager.SetProperty(nextTrack, x => x.End, selectedTrack.End);
+            //    }
+            //    var newPosition = (ushort?)(selectedTrack.Position + 1);
+            //    _trackManager.SetProperty(selectedTrack, x => x.Position, newPosition);
+            //    _trackManager.SetProperty(selectedTrack, x => x.Begin, newBegin);
+            //    _trackManager.SetProperty(selectedTrack, x => x.End, newEnd);
+            //}
+            //SetValue(cuesheet!, x => x.Tracks, cuesheet?.Tracks.OrderBy(x => x.Position));
+            //_traceChangeManager.BulkEdit = false;
             return Result.Success();
         }
 
@@ -272,11 +221,13 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
 
         static Track? GetLastTrack(Cuesheet cuesheet)
         {
-            return cuesheet.Tracks
-                .OrderByDescending(x => x.Position.HasValue).ThenBy(x => x.Position)
-                .ThenByDescending(x => x.Begin.HasValue).ThenBy(x => x.Begin)
-                .ThenByDescending(x => x.End.HasValue).ThenBy(x => x.End)
-                .LastOrDefault();
+            //TODO
+            return null;
+            //return cuesheet.Tracks
+            //    .OrderByDescending(x => x.Position.HasValue).ThenBy(x => x.Position)
+            //    .ThenByDescending(x => x.Begin.HasValue).ThenBy(x => x.Begin)
+            //    .ThenByDescending(x => x.End.HasValue).ThenBy(x => x.End)
+            //    .LastOrDefault();
         }
     }
 }
