+++ AudioCuesheetEditor/Services/IO/CuesheetExportService.cs
@@ -38,7 +38,8 @@ namespace AudioCuesheetEditor.Services.IO
                 validationMessages.Add(new ValidationMessage("File extension is not '{0}'", FileExtensions.Cuesheet));
             }
             validationMessages.AddRange(_sessionStateContainer.Cuesheet.Validate().ValidationMessages);
-            validationMessages.AddRange(_sessionStateContainer.Cuesheet.Tracks.Select(x => x.Validate()).SelectMany(x => x.ValidationMessages));
+            //TODO
+            //validationMessages.AddRange(_sessionStateContainer.Cuesheet.Tracks.Select(x => x.Validate()).SelectMany(x => x.ValidationMessages));
             if (validationMessages.Count != 0)
             {
                 return Result.Failure(new Error(ErrorType.ValidationFailed, string.Join(Environment.NewLine, validationMessages.Select(x => x.GetMessageLocalized(_localizer)))));
@@ -82,39 +83,39 @@ namespace AudioCuesheetEditor.Services.IO
             builder.AppendLine(string.Format("{0} \"{1}\"", CuesheetConstants.CuesheetArtist, _sessionStateContainer.Cuesheet.Artist));
             //TODO
             //builder.AppendLine(string.Format("{0} \"{1}\" {2}", CuesheetConstants.CuesheetFileName, audiofileName, _sessionStateContainer.Cuesheet.Audiofile?.AudioFileType));
-            IEnumerable<Track> tracks = _sessionStateContainer.Cuesheet.Tracks.OrderBy(x => x.Position);
-            if (tracks.Any())
-            {
-                //Position and begin should always start from 0 even with splitpoints
-                int positionDifference = 1 - Convert.ToInt32(tracks.First().Position);
-                foreach (var track in tracks)
-                {
-                    builder.AppendLine(string.Format("{0}{1} {2:00} {3}", CuesheetConstants.Tab, CuesheetConstants.CuesheetTrack, track.Position + positionDifference, CuesheetConstants.CuesheetTrackAudio));
-                    builder.AppendLine(string.Format("{0}{1}{2} \"{3}\"", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackTitle, track.Title));
-                    builder.AppendLine(string.Format("{0}{1}{2} \"{3}\"", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackArtist, track.Artist));
-                    if (track.Flags.Any())
-                    {
-                        builder.AppendLine(string.Format("{0}{1}{2} {3}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackFlags, string.Join(" ", track.Flags.Select(x => x.CuesheetLabel))));
-                    }
-                    if (track.PreGap.HasValue)
-                    {
-                        builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackPreGap, Math.Floor(track.PreGap.Value.TotalMinutes), track.PreGap.Value.Seconds, track.PreGap.Value.Milliseconds * 75 / 1000));
-                    }
-                    if (track.Begin.HasValue)
-                    {
-                        var begin = track.Begin.Value;
-                        builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackIndex01, Math.Floor(begin.TotalMinutes), begin.Seconds, begin.Milliseconds * 75 / 1000));
-                    }
-                    else
-                    {
-                        throw new NullReferenceException(string.Format("{0} may not be null!", nameof(Track.Begin)));
-                    }
-                    if (track.PostGap.HasValue)
-                    {
-                        builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackPostGap, Math.Floor(track.PostGap.Value.TotalMinutes), track.PostGap.Value.Seconds, track.PostGap.Value.Milliseconds * 75 / 1000));
-                    }
-                }
-            }
+            //IEnumerable<Track> tracks = _sessionStateContainer.Cuesheet.Tracks.OrderBy(x => x.Position);
+            //if (tracks.Any())
+            //{
+            //    //Position and begin should always start from 0 even with splitpoints
+            //    int positionDifference = 1 - Convert.ToInt32(tracks.First().Position);
+            //    foreach (var track in tracks)
+            //    {
+            //        builder.AppendLine(string.Format("{0}{1} {2:00} {3}", CuesheetConstants.Tab, CuesheetConstants.CuesheetTrack, track.Position + positionDifference, CuesheetConstants.CuesheetTrackAudio));
+            //        builder.AppendLine(string.Format("{0}{1}{2} \"{3}\"", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackTitle, track.Title));
+            //        builder.AppendLine(string.Format("{0}{1}{2} \"{3}\"", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackArtist, track.Artist));
+            //        if (track.Flags.Any())
+            //        {
+            //            builder.AppendLine(string.Format("{0}{1}{2} {3}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackFlags, string.Join(" ", track.Flags.Select(x => x.CuesheetLabel))));
+            //        }
+            //        if (track.PreGap.HasValue)
+            //        {
+            //            builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackPreGap, Math.Floor(track.PreGap.Value.TotalMinutes), track.PreGap.Value.Seconds, track.PreGap.Value.Milliseconds * 75 / 1000));
+            //        }
+            //        if (track.Begin.HasValue)
+            //        {
+            //            var begin = track.Begin.Value;
+            //            builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackIndex01, Math.Floor(begin.TotalMinutes), begin.Seconds, begin.Milliseconds * 75 / 1000));
+            //        }
+            //        else
+            //        {
+            //            throw new NullReferenceException(string.Format("{0} may not be null!", nameof(Track.Begin)));
+            //        }
+            //        if (track.PostGap.HasValue)
+            //        {
+            //            builder.AppendLine(string.Format("{0}{1}{2} {3:00}:{4:00}:{5:00}", CuesheetConstants.Tab, CuesheetConstants.Tab, CuesheetConstants.TrackPostGap, Math.Floor(track.PostGap.Value.TotalMinutes), track.PostGap.Value.Seconds, track.PostGap.Value.Milliseconds * 75 / 1000));
+            //        }
+            //    }
+            //}
             return builder.ToString();
         }
     }
