+++ AudioCuesheetEditor/Shared/TrackList/TrackList.razor
@@ -29,12 +29,13 @@ along with Foobar.  If not, see
 @inject ITrackManager _trackManager
 @inject ICuesheetManager _cuesheetManager
 @inject ITraceChangeManager _traceChangeManager
+@inject IAudiofileManager _audiofileManager
 
 @switch (CurrentViewMode)
 {
     case ViewMode.DetailView:
     case ViewMode.ImportView:
-        var validationResult = Cuesheet?.Validate(nameof(Cuesheet.Tracks));
+        var validationResult = Audiofile.Validate(nameof(Audiofile.Tracks));
         if (validationResult?.Status == ValidationStatus.Error)
         {
             <MudAlert Severity="Severity.Error" Variant="Variant.Filled">
@@ -48,7 +49,7 @@ along with Foobar.  If not, see
         <div aria-label="@_localizer["Track table controls"]">
             <TrackListControlButtons AddButtonClicked="AddTrackClicked" EditSelectedTracksClicked="EditSelectedTracksClicked" EditSelectedTracksDisabled="_selectedTracks.Count == 0"
                                      DeleteSelectedTracksClicked="DeleteSelectedTracksClicked" DeleteSelectedTracksDisabled="_selectedTracks.Count == 0"
-                                     DeleteAllTracksClicked="OnDeleteAllTracksClicked" DeleteAllTracksDisabled="Cuesheet?.Tracks.Any() == false"
+                                     DeleteAllTracksClicked="OnDeleteAllTracksClicked" DeleteAllTracksDisabled="Audiofile.Tracks.Any() == false"
                                      MoveTracksUpDisabled="!_cuesheetManager.IsMoveTracksUpPossible(_selectedTracks)" MoveTracksUpClicked="() => _cuesheetManager.MoveTracksUp(_selectedTracks)"
                                      MoveTracksDownDisabled="!_cuesheetManager.IsMoveTracksDownPossible(_selectedTracks)" MoveTracksDownClicked="(MoveTracksDownClicked) => _cuesheetManager.MoveTracksDown(_selectedTracks)"
                                      CopySelectedTracksDisabled="_selectedTracks.Count != 1" CopySelectedTracksClicked="() => CopyTrackClicked()"
@@ -58,7 +59,7 @@ along with Foobar.  If not, see
 }
 
 <MudForm @ref="_form">
-    <MudDataGrid T="Track" Items="Cuesheet?.Tracks" ReadOnly="false" Bordered MultiSelection SelectOnRowClick="false" @bind-SelectedItems="_selectedTracks" EditTrigger="DataGridEditTrigger.OnRowClick"
+    <MudDataGrid T="Track" Items="Audiofile.Tracks" ReadOnly="false" Bordered MultiSelection SelectOnRowClick="false" @bind-SelectedItems="_selectedTracks" EditTrigger="DataGridEditTrigger.OnRowClick"
                  EditMode="DataGridEditMode.Cell" ColumnResizeMode="ResizeMode.Column" RowContextMenuClick="OpenMenuContent" Height="@(_applicationOptions?.FixedTracksTableHeader == true ? "600px" : null)" 
                  FixedHeader="_applicationOptions?.FixedTracksTableHeader == true" Validator="_form" Virtualize SortMode="SortMode.None">
         <Columns>
@@ -108,7 +109,7 @@ along with Foobar.  If not, see
             </PropertyColumn>
             <TemplateColumn Title="@_localizer["Status"]" Hidden="CurrentViewMode != ViewMode.DetailView" HeaderStyle="width: 1px;">
                 <EditTemplate>
-                    @if (Cuesheet?.Tracks.FirstOrDefault() != context.Item)
+                    @if (Audiofile.Tracks.FirstOrDefault() != context.Item)
                     {
                         <MudTooltip Text="@GetLinkedTrackTooltip(context.Item)">
                             <MudChip Icon="@GetLinkedTrackIcon(context.Item)" Color="Color.Secondary" OnClick="() => _trackManager.SetProperty(context.Item, x => x.IsLinkedToPreviousTrack, !context.Item.IsLinkedToPreviousTrack)" />
@@ -128,24 +129,17 @@ along with Foobar.  If not, see
     }
 </MudMenu>
 
-@code {
+    @code {
     HashSet<Track> _selectedTracks = new();
     MudMenu? _trackContextMenu;
     Track? _contextMenuTrack;
     ApplicationOptions? _applicationOptions;
     MudForm? _form;
+    Cuesheet? _cuesheet => _sessionStateContainer.GetActiveCuesheet();
 
-    public Cuesheet? Cuesheet
-    {
-        get
-        {
-            if (CurrentViewMode == ViewMode.ImportView)
-            {
-                return _sessionStateContainer.ImportCuesheet;
-            }
-            return _sessionStateContainer.Cuesheet;
-        }
-    }
+    [Parameter]
+    [EditorRequired]
+    public Audiofile Audiofile { get; set; }
 
     [CascadingParameter]
     public ViewMode CurrentViewMode { get; set; }
@@ -173,7 +167,7 @@ along with Foobar.  If not, see
         {
             IsLinkedToPreviousTrack = _applicationOptions!.DefaultIsLinkedToPreviousTrack
         };
-        _cuesheetManager.AddTrack(newTrack);
+        _audiofileManager.AddTrack(Audiofile, newTrack);
     }
 
     async Task EditSelectedTracksClicked()
@@ -197,8 +191,9 @@ along with Foobar.  If not, see
         var result = await dialog.Result;
         if (result?.Canceled == false)
         {
-            _cuesheetManager.RemoveTracks(Cuesheet!.Tracks);
-            _selectedTracks.Clear();
+            //TODO: Switch to Audiofile
+            // _cuesheetManager.RemoveTracks(_cuesheet!.Tracks);
+            // _selectedTracks.Clear();
         }
     }
 
@@ -234,7 +229,7 @@ along with Foobar.  If not, see
             copy.Position = (ushort?)(trackThatWillBeCopied.Position + 1);
             copy.Begin = trackThatWillBeCopied.End;
             copy.End = null;
-            _cuesheetManager.AddTrack(copy);
+            _audiofileManager.AddTrack(Audiofile, copy);
             await EditTrackModalClicked(copy);
         }
     }
